added calc_days to lib, parsing stays in main, tests moved to lib
